fix(ui): change component names from node to action and refactor
