+++ ui/src/components/ActionItem/index.tsx
@@ -1,5 +1,6 @@
 import { forwardRef } from "react";
 
+import { typeColorClass } from "@flow/features/Editor/components/OverlayUI/components/ActionPickerDialog/utils";
 import { cn } from "@flow/lib/utils";
 import { Action } from "@flow/types";
 import { getNodeIcon } from "@flow/utils/getNodeIcon";
@@ -12,19 +13,6 @@ type Props = {
   isSelected?: boolean;
   onSingleClick?: (actionName: string) => void;
   onDoubleClick?: (actionName: string) => void;
-};
-
-const typeColorClass = (type: string) => {
-  switch (type) {
-    case "transformer":
-      return "bg-node-transformer/80";
-    case "reader":
-      return "bg-node-reader/80";
-    case "writer":
-      return "bg-node-writer/80";
-    default:
-      return "bg-secondary";
-  }
 };
 
 const ActionItem = forwardRef<HTMLDivElement, Props>(
@@ -61,11 +49,11 @@ const ActionItem = forwardRef<HTMLDivElement, Props>(
             <Icon size={12} weight="thin" className="text-white" />
           </div>
           <span className="flex-1 truncate text-sm">{action.name}</span>
-          {/* <span className="shrink-0 text-xs capitalize opacity-60">
-                              {action.type}
-                            </span> */}
           <div
-            className={`self-center rounded border  ${action.type === "transformer" ? "bg-node-transformer/95 dark:bg-node-transformer/60" : action.type === "reader" ? "bg-node-reader/95 dark:bg-node-reader/60" : action.type === "writer" ? "bg-node-writer/85 dark:bg-node-writer/30" : "bg-popover"} p-0.5 align-middle`}>
+            className={cn(
+              "self-center rounded border p-1 align-middle",
+              typeColorClass(action.type),
+            )}>
             <p className="self-center text-xs text-zinc-200 capitalize">
               {action.type}
             </p>
