+++ ui/src/features/Editor/components/OverlayUI/components/index.ts
@@ -1,6 +1,6 @@
 export { default as CanvasActionBar } from "./CanvasActionbar";
 export { default as Toolbox } from "./Toolbox";
-export { default as NodePickerDialog } from "./NodePickerDialog";
+export { default as NodePickerDialog } from "./ActionPickerDialog";
 export { default as LayoutOptionsDialog } from "./LayoutOptionsDialog";
 export { default as DebugLogs } from "./DebugPanel/DebugLogs";
 export { default as DebugPreview } from "./DebugPanel/DebugPreview";
